+++ src/xSTIR/cSTIR/cstir.cpp
@@ -78,6 +78,8 @@ void* cSTIR_newObject(const char* name)
 			return NEW_OBJECT_HANDLE(AcqModUsingMatrix3DF);
 		if (boost::iequals(name, "RayTracingMatrix"))
 			return NEW_OBJECT_HANDLE(RayTracingMatrix);
+		if (boost::iequals(name, "SPECTUBMatrix"))
+			return NEW_OBJECT_HANDLE(SPECTUBMatrix);
 		if (boost::iequals(name, "QuadraticPrior"))
 			return NEW_OBJECT_HANDLE(QuadPrior3DF);
 		if (boost::iequals(name, "PLSPrior"))
@@ -113,6 +115,8 @@ void* cSTIR_setParameter
 			return cSTIR_setAcqModUsingMatrixParameter(hs, name, hv);
 		else if (boost::iequals(obj, "RayTracingMatrix"))
 			return cSTIR_setRayTracingMatrixParameter(hs, name, hv);
+		else if (boost::iequals(obj, "SPECTUBMatrix"))
+			return cSTIR_setSPECTUBMatrixParameter(hs, name, hv);
 		else if (boost::iequals(obj, "GeneralisedPrior"))
 			return cSTIR_setGeneralisedPriorParameter(hs, name, hv);
 		else if (boost::iequals(obj, "QuadraticPrior"))
@@ -159,6 +163,8 @@ void* cSTIR_parameter(const void* ptr, const char* obj, const char* name)
 			(handle, name);
 		else if (boost::iequals(obj, "RayTracingMatrix"))
 			return cSTIR_rayTracingMatrixParameter(handle, name);
+		else if (boost::iequals(obj, "SPECTUBMatrix"))
+			return cSTIR_SPECTUBMatrixParameter(handle, name);
 		else if (boost::iequals(obj, "AcqModUsingMatrix"))
 			return cSTIR_acqModUsingMatrixParameter(handle, name);
 		else if (boost::iequals(obj, "GeneralisedPrior"))