+++ src/xSTIR/cSTIR/include/sirf/cSTIR/stir_x.h
@@ -1,6 +1,7 @@
 /*
 CCP PETMR Synergistic Image Reconstruction Framework (SIRF)
 Copyright 2015 - 2017 Rutherford Appleton Laboratory STFC
+Copyright 2019 University College London
 
 This is software developed for the Collaborative Computational
 Project in Positron Emission Tomography and Magnetic Resonance imaging
@@ -33,9 +34,7 @@ limitations under the License.
 */
 
 #include <stdlib.h>
-
 #include "sirf/cSTIR/stir_data_containers.h"
-
 #define MIN_BIN_EFFICIENCY 1.0e-20f
 //#define MIN_BIN_EFFICIENCY 1.0e-6f