+++ CHANGES.md
@@ -8,6 +8,7 @@
 * implemented sorting of MR images
 * implemented reading of MR acquisition data from ISMRMRD file
 * Matlab: in keeping with changes to c++ and python, classes are now called with e.g., `sirf.STIR.obj` instead of `mSTIR.obj`. Aliases can be used to shorten this (e.g., `PET=set_up_PET()` and then `PET.obj`).
+* added SPECTUBMatrix for (simple) usage in SPECT
 
 ## v1.1.0